requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.4.6
aiohttp==3.9.5
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.1.0
bitarray==3.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
ckzg==2.0.1
colorama==0.4.6
customtkinter==5.2.2
cytoolz==1.0.1
darkdetect==0.8.0
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
frozenlist==1.5.0
greenlet==3.1.1
hexbytes==1.3.0
idna==3.10
loguru==0.7.3
magic-filter==1.0.12
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
packaging==24.2
parsimonious==0.10.0
patchright==1.50.0
primp==0.14.0
prompt_toolkit==3.0.50
propcache==0.3.0
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
Pygments==2.19.1
pytz==2025.1
pyunormalize==16.0.0
PyYAML==6.0.2
questionary==2.1.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rlp==4.1.0
tabulate==0.9.0
toolz==1.0.0
types-requests==2.32.0.20250301
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
web3==7.8.0
websockets==13.1
win32_setctime==1.2.0
yarl==1.18.3
ccxt
pynocaptcha
curl_cffi
tqdm
pandas
openpyxl
